+++ src/pages/Auth/Signin.jsx
@@ -7,7 +7,7 @@ import sketchLogo from "../../assets/logo1.png";
 import { InputField } from "../../components";
 
 import { testCredentials, initialFormValues } from "./helper";
-import { loginUser } from "../../services/authentication";
+import { loginUser } from "../../services";
 import { useAuthContext, useGlobalContext } from "../../context";
 import useScrollToTop from "../../hooks/useScrollToTop";
