+++ src/pages/Auth/Signup.jsx
@@ -8,7 +8,7 @@ import { InputField } from "../../components";
 
 import { initialFormValues, validateInputs } from "./helper";
 
-import { registerUser } from "../../services/authentication";
+import { registerUser } from "../../services";
 import { useAuthContext, useGlobalContext } from "../../context";
 import useScrollToTop from "../../hooks/useScrollToTop";
