+++ src/pages/History/History.jsx
@@ -2,7 +2,7 @@ import { useDataContext, useGlobalContext } from "../../context";
 
 import EmptyState from "../EmptyState/EmptyState";
 import { Loader, HorizontalCard } from "../../components";
-import { clearHistory } from "../../services/history";
+import { clearHistory } from "../../services";
 import useScrollToTop from "../../hooks/useScrollToTop";
 
 import "./history.scss";
